+++ src/projectManager/projectManagerViewProvider.ts
@@ -18,7 +18,7 @@ import {
     openProject,
     createNewProject,
 } from "../utils/projectCreationUtils/projectCreationUtils";
-import { FrontierAPI } from "webviews/codex-webviews/src/StartupFLow/types";
+import { FrontierAPI } from "webviews/codex-webviews/src/StartupFlow/types";
 import { waitForExtensionActivation } from "../utils/vscode";
 import git from "isomorphic-git";
 import * as fs from "fs";
