+++ src/providers/StartupFlow/StartupFlowProvider.ts
@@ -10,7 +10,7 @@ import {
 import * as vscode from "vscode";
 import { PreflightCheck, PreflightState } from "./preflight";
 import { findAllCodexProjects } from "../../../src/projectManager/utils/projectUtils";
-import { AuthState, FrontierAPI } from "webviews/codex-webviews/src/StartupFLow/types";
+import { AuthState, FrontierAPI } from "webviews/codex-webviews/src/StartupFlow/types";
 import { CustomWebviewProvider } from "../../projectManager/projectManagerViewProvider";
 import {
     createNewProject,
