+++ webviews/codex-webviews/src/StartupFlow/components/LoginRegisterStep.tsx
