+++ webviews/codex-webviews/src/StartupFlow/components/WorkspaceStep.tsx
